refa: Move <Label> and <HelpText> to components root
- remove legacy label component
- no docs for <HelpText> since we don't know if we want to expose it
