--- docs/content/components/validation-message.mdx
@@ -1,41 +1,0 @@
----
-title: ValidationMessage
----
-
-import { FigmaLink } from '../../src/components/FigmaLink';
-
-# ValidationMessage
-
-Simple colored element to display a validation message as a reaction to an event.
-
-<MarigoldTheme>
-  <ValidationMessage>
-    <Exclamation />
-    Validation message
-  </ValidationMessage>
-</MarigoldTheme>
-
-```tsx onlyCode
-import { ValidationMessage } from '@marigold/components';
-```
-
-<FigmaLink
-  to={
-    'https://www.figma.com/file/DFKyTGHAoDxOsUBPszLLxP/%F0%9F%8F%B5%EF%B8%8FMarigold?node-id=471%3A3'
-  }
-/>
-
-## Props
-
-| Property             | Type     | Default |
-| :------------------- | :------- | :------ |
-| `variant` (optional) | `string` | `error` |
-
-## Examples
-
-```tsx expandCode
-<ValidationMessage>
-  <Exclamation />
-  Validation message
-</ValidationMessage>
-```
