+++ packages/components/src/HelpText/index.ts
@@ -1,0 +1,1 @@
+export * from './HelpText';
