+++ packages/components/src/Label/Label.stories.tsx
@@ -5,37 +5,23 @@ import { Label } from './Label';
 export default {
   title: 'Components/Label',
   argTypes: {
-    variant: {
+    children: {
       control: {
         type: 'text',
       },
-      description: '?',
-      table: {
-        defaultValue: {
-          summary: 'above',
-        },
-      },
-    },
-    htmlFor: {
-      control: {
-        type: 'text',
-      },
-      defaultValue: 'input',
+      description: 'Text of the label',
+      defaultValue: 'Label',
     },
     required: {
       control: {
         type: 'boolean',
       },
-      description: 'Adds required Icon to label',
-      table: {
-        defaultValue: {
-          summary: false,
-        },
-      },
+      description: 'Hint that the related control required',
+      defaultValue: false,
     },
   },
 } as Meta;
 
-export const Basic: ComponentStory<typeof Label> = args => (
-  <Label {...args}>Label</Label>
+export const Basic: ComponentStory<typeof Label> = ({ children, ...args }) => (
+  <Label {...args}>{children}</Label>
 );
